Rewrite theme based on CSS flexboxes
The theme looks pretty much the same, but with CSS flexboxes it's possible
to ensure that wide segments (like the window title) won't push other content
outside of the statusline. The gradient mask also works much better since there's
no need to set a max width on the window title and later on the notification
daemon notifications.

Closes #3